[xcb] ignore non-generated configure notify events
There seems to be little to no documentation on this, however testing
shows that the non-generated configure notify events contain
invalid/incorrect x&y details when a window is moved. xtrace shows this
also and checks for bit 0x80 calling it `generated`. So to fix this we
ignore the non-generated configure notify messages.